requirements.txt
streamlit
requests
huggingface_hub
fpdf2
